RavenDB-26861 Add Open database link
